+++ src/main/java/com/myspring/spring/qna/QnaService.java
@@ -115,8 +115,10 @@ public class QnaService {
 
 			file = new File("./images/qna/" + requestData.getQnaNo() + "/");
 			underDir = file.listFiles();
-			for (int i = 0; i < underDir.length; i++) {
-				underDir[i].delete();
+			if (underDir != null) {
+				for (int i = 0; i < underDir.length; i++) {
+					underDir[i].delete();
+				}
 			}
 			if (fileList != null) {
 				for (int i = 0; i < fileList.size(); i++) {
@@ -171,8 +173,10 @@ public class QnaService {
 
 			file = new File("./images/qna/" + qnaNo + "/");
 			underDir = file.listFiles();
-			for (int i = 0; i < underDir.length; i++) {
-				underDir[i].delete();
+			if (underDir != null) {
+				for (int i = 0; i < underDir.length; i++) {
+					underDir[i].delete();
+				}
 			}
 			entity = new ResponseEntity<>(HttpStatus.OK);
 		} catch (Exception e) {
