+++ src/main/java/com/myspring/spring/review/ReviewService.java
@@ -88,8 +88,10 @@ public class ReviewService {
 				// 폴더 내 모든 파일 삭제
 				file = new File("./images/review/" + reviewNo + "/");
 				underDir = file.listFiles();
-				for (int i = 0; i < underDir.length; i++) {
-					underDir[i].delete();
+				if (underDir != null) {
+					for (int i = 0; i < underDir.length; i++) {
+						underDir[i].delete();
+					}
 				}
 				entity = new ResponseEntity<>(HttpStatus.OK);
 			} catch (Exception e) {
@@ -114,8 +116,10 @@ public class ReviewService {
 			// 폴더 내 모든 파일 삭제
 			file = new File("./images/review/" + requestData.getReviewNo() + "/");
 			underDir = file.listFiles();
-			for (int i = 0; i < underDir.length; i++) {
-				underDir[i].delete();
+			if (underDir != null) {
+				for (int i = 0; i < underDir.length; i++) {
+					underDir[i].delete();
+				}
 			}
 			if (fileList != null) {
 				MultipartFile multipartFile = fileList.get(0);
